Extract access control logic
