+++ cmd/catalyst-node/catalyst-node.go
@@ -1,8 +1,6 @@
 package main
 
 import (
-	"bytes"
-	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"flag"
@@ -17,13 +15,11 @@ import (
 	"os/signal"
 	"regexp"
 	"runtime"
-	"strconv"
 	"strings"
 	"sync"
 	"syscall"
 	"time"
 
-	"github.com/golang-jwt/jwt/v4"
 	serfclient "github.com/hashicorp/serf/client"
 	"github.com/hashicorp/serf/cmd/serf/command/agent"
 	"github.com/livepeer/livepeer-data/pkg/mistconnector"
@@ -759,229 +755,4 @@ func queryMistForClosestNodeSource(playbackID, lat, lon, prefix string, source b
 		return "", fmt.Errorf("GET request '%s' returned 'FULL'", murl)
 	}
 	return string(body), nil
-}
-
-type PlaybackAccessControl struct {
-	gateURL string
-	cache   map[string]map[string]*PlaybackAccessControlEntry
-	*http.Client
-}
-
-type PlaybackAccessControlEntry struct {
-	Stale  time.Time
-	MaxAge time.Time
-	allow  bool
-}
-
-type PlaybackAccessControlRequest struct {
-	Type   string `json:"type"`
-	Pub    string `json:"pub"`
-	Stream string `json:"stream"`
-}
-
-const UserNewTrigger = "USER_NEW"
-
-func triggerHandler(gateURL string) http.Handler {
-	playbackAccessControl := PlaybackAccessControl{
-		gateURL,
-		make(map[string]map[string]*PlaybackAccessControlEntry),
-		&http.Client{},
-	}
-
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		payload, err := io.ReadAll(r.Body)
-
-		if err != nil {
-			w.Write([]byte("false"))
-			glog.Errorf("Unable to parse trigger body %v", err)
-			return
-		}
-
-		triggerName := r.Header.Get("X-Trigger")
-
-		switch triggerName {
-		case UserNewTrigger:
-			w.Write(handleUserNew(&playbackAccessControl, payload))
-			return
-		default:
-			w.Write([]byte("false"))
-			glog.Errorf("Trigger not handled %v", triggerName)
-			return
-		}
-
-	})
-}
-
-func handleUserNew(ac *PlaybackAccessControl, payload []byte) []byte {
-	lines := strings.Split(string(payload), "\n")
-
-	if len(lines) != 6 {
-		glog.Errorf("Malformed trigger payload")
-		return []byte("false")
-	}
-
-	requestURL, err := url.Parse(lines[4])
-	if err != nil {
-		glog.Errorf("Unable to parse URL %v", err)
-		return []byte("false")
-	}
-
-	playbackID := lines[0]
-	jwtToken := requestURL.Query().Get("jwt")
-
-	var pubKey string
-	if jwtToken != "" {
-		claims, err := decodeJwt(jwtToken)
-		if err != nil {
-			glog.Errorf("Unable to decode JWT token %v", err)
-			return []byte("false")
-		}
-
-		if err := claims.Valid(); err != nil {
-			glog.Errorf("Invalid token %v", err)
-			return []byte("false")
-		}
-
-		if playbackID != claims.Subject {
-			glog.Errorf("PlaybackId mismatch")
-			return []byte("false")
-		}
-
-		pubKey = claims.PublicKey
-
-		if time.Now().After(claims.ExpiresAt.Time) {
-			glog.Errorf("Token expired %v", claims.ExpiresAt)
-			return []byte("false")
-		}
-	}
-
-	playbackAccessControl, err := getPlaybackAccessControlInfo(ac, playbackID, pubKey)
-	if err != nil {
-		glog.Errorf("Unable to get playback access control info %v", err)
-		return []byte("false")
-	}
-
-	if playbackAccessControl[pubKey].allow {
-		return []byte("true")
-	}
-
-	return []byte("false")
-}
-
-func getPlaybackAccessControlInfo(ac *PlaybackAccessControl, playbackID, pubKey string) (map[string]*PlaybackAccessControlEntry, error) {
-	if ac.cache[playbackID] == nil ||
-		ac.cache[playbackID][pubKey] == nil ||
-		time.Now().After(ac.cache[playbackID][pubKey].Stale) {
-
-		err := cachePlaybackAccessControlInfo(ac, playbackID, pubKey)
-		if err != nil {
-			return nil, err
-		}
-
-	} else if time.Now().After(ac.cache[playbackID][pubKey].MaxAge) {
-		go cachePlaybackAccessControlInfo(ac, playbackID, pubKey)
-	}
-
-	return ac.cache[playbackID], nil
-}
-
-func cachePlaybackAccessControlInfo(ac *PlaybackAccessControl, playbackID, pubKey string) error {
-	body, err := json.Marshal(PlaybackAccessControlRequest{"jwt", pubKey, playbackID})
-	if err != nil {
-		return err
-	}
-
-	statusCode, maxAge, stale, err := queryGate(ac, body)
-	if err != nil {
-		return err
-	}
-
-	var allow = statusCode == 204
-	var maxAgeTime = time.Now().Add(time.Duration(maxAge) * time.Second)
-	var staleTime = time.Now().Add(time.Duration(stale) * time.Second)
-
-	if ac.cache[playbackID] == nil {
-		ac.cache[playbackID] = make(map[string]*PlaybackAccessControlEntry)
-	}
-
-	ac.cache[playbackID][pubKey] = &PlaybackAccessControlEntry{staleTime, maxAgeTime, allow}
-
-	return nil
-}
-
-var queryGate = func(ac *PlaybackAccessControl, body []byte) (int, int64, int64, error) {
-	req, err := http.NewRequest("POST", ac.gateURL, bytes.NewReader(body))
-	if err != nil {
-		return 0, 0, 0, err
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-
-	res, err := ac.Client.Do(req)
-	if err != nil {
-		return 0, 0, 0, err
-	}
-
-	defer res.Body.Close()
-
-	cacheControlDirectives := strings.Split(res.Header.Get("Cache-Control"), ",")
-
-	var maxAge, stale int64 = 0, 0
-	if len(cacheControlDirectives) == 2 {
-		maxAge, err = strconv.ParseInt(strings.Split(cacheControlDirectives[0], "=")[1], 10, 64)
-		if err != nil {
-			return 0, 0, 0, err
-		}
-
-		stale, err = strconv.ParseInt(strings.Split(cacheControlDirectives[1], "=")[1], 10, 64)
-		if err != nil {
-			return 0, 0, 0, err
-		}
-	}
-
-	return res.StatusCode, maxAge, stale, nil
-}
-
-type PlaybackGateClaims struct {
-	PublicKey string `json:"pub"`
-	*jwt.RegisteredClaims
-}
-
-func (c *PlaybackGateClaims) Valid() error {
-	if err := c.RegisteredClaims.Valid(); err != nil {
-		return err
-	}
-	if c.Subject == "" {
-		return errors.New("missing sub claim")
-	}
-	if c.PublicKey == "" {
-		return errors.New("missing pub claim")
-	}
-	if c.ExpiresAt == nil {
-		return errors.New("missing exp claim")
-	} else if time.Until(c.ExpiresAt.Time) > 7*24*time.Hour {
-		return errors.New("exp claim too far in the future")
-	}
-	return nil
-}
-
-func decodeJwt(tokenString string) (*PlaybackGateClaims, error) {
-	token, err := jwt.ParseWithClaims(tokenString, &PlaybackGateClaims{}, func(token *jwt.Token) (interface{}, error) {
-		pub := token.Claims.(*PlaybackGateClaims).PublicKey
-		decodedPubkey, err := base64.StdEncoding.DecodeString(pub)
-		if err != nil {
-			return nil, err
-		}
-
-		return jwt.ParseECPublicKeyFromPEM(decodedPubkey)
-	})
-
-	if err != nil {
-		return nil, err
-	} else if err = token.Claims.Valid(); err != nil {
-		return nil, err
-	} else if !token.Valid {
-		return nil, errors.New("invalid token")
-	}
-	return token.Claims.(*PlaybackGateClaims), nil
 }
